chore(refactor): Add helpers and configuration files
- Add reused functions into helpers.js file
- Centralized important variables and settings to config.js file

[Finishes LEC 293]
