+++ src/js/config.js
@@ -1,0 +1,7 @@
+// For all the variables we will reuse throughout the project
+// Responsible for defining important data
+
+// UPPERCASE for constants
+// 'https://forkify-api.herokuapp.com/api/v2/recipes/5ed6604591c37cdc054bc886'
+export const API_URL = `https://forkify-api.herokuapp.com/api/v2/recipes`;
+export const TIMEOUT_SEC = 10;
